[WIP] Lowering Ion dialect to llvm dialect (#1466)
**Context:**
So far, as part of the ion dialect pass infrastructure, we have the ion
decomposition pass which genreates a universal gate set of (RX,RY, MS).
This gate set would be lowered to ion dialect which consists of IonOp,
PulseOp and ParallelProtocolOp operations and IonType. we need to lower
the ion dialect operations into llvm stubs which will be used in runtime
to generate machine related instructions.

**Description of the Change:**
- [x] Added a new lowering pass `--convert-ion-to-llvm`
- [x] Added IonOp lowering to @__catalyst_ion 
- [x] Added IonType lowering
- [x] Added pulseOp lowering to @ __catalyst_pulse_op 
- [x] Added parallelprotocolOp lowering to @
__catalyst_parallel_protocol
- [x] Added changelog

**Benefits:**
This PR would enable runtime to capture the ion dialect related
instructions.

**Possible Drawbacks:**

**Related GitHub Issues:**
[sc-79342]

---------

+++ mlir/include/Ion/Transforms/oqd_database_types.hpp
@@ -95,11 +95,11 @@ struct Ion {
     // It contains the innate properties of the qubit.
     std::string name;
     double mass, charge;
-    std::vector<int64_t> position;
+    std::vector<double> position;
     std::vector<Level> levels;
     std::vector<Transition> transitions;
 
-    Ion(std::string _name, double _mass, double _charge, std::vector<int64_t> _position,
+    Ion(std::string _name, double _mass, double _charge, std::vector<double> _position,
         std::vector<Level> _levels, std::vector<Transition> _transitions)
         : name(_name), mass(_mass), charge(_charge), position(_position), levels(_levels),
           transitions(_transitions)
